Changes from today: setup navigation with login and home screens
- Added nav_graph and NavHostFragment for stable navigation.
- Built LoginFragment and HomeFragment with basic UI.
- Configured MainActivity with NavController and toolbar logo.
