+++ app/src/main/java/com/example/dormconnectapp/Noticeboardfragment.kt
@@ -1,0 +1,27 @@
+package com.example.dormconnectapp
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.dormconnectapp.databinding.FragmentNoticeBoardBinding
+
+class NoticeBoardFragment : Fragment() {
+
+    private var _binding: FragmentNoticeBoardBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentNoticeBoardBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
